pyproject.toml
[build-system]
requires = ["setuptools>=72", "setuptools-scm>=8"]
# PEP 517 standard backend (editable installs). The legacy backend was removed
# in newer setuptools and breaks `pip install -e .` on CI runners.
build-backend = "setuptools.build_meta"

[project]
name = "alignai"
dynamic = ["version"]
description = "AI-powered resume and cover-letter aligner with Telegram bot interface"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
keywords = ["resume", "cover-letter", "ai", "agents", "openai"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    # Agents & LLM
    "openai-agents>=0.0.12",
    "openai>=1.30",
    # UI
    "PySide6>=6.7",
    # HTTP & scraping
    "httpx>=0.27",
    "trafilatura>=1.12",
    "playwright>=1.44",
    # Documents
    "pypdf>=4.3",
    "python-docx>=1.1",
    "jinja2>=3.1",
    # Storage & settings
    "platformdirs>=4.2",
    "keyring>=25.2",
    # Telegram
    "python-telegram-bot>=21.3",
    # Utilities
    "pydantic>=2.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "import-linter>=2.1",
    "types-Pygments",
]

packaging = [
    "pyinstaller>=6.0",
]

[project.scripts]
alignai = "alignai.main:main"

[tool.setuptools_scm]
# Version is derived from git tags (e.g. v1.2.3); no manual edits needed.
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"

[tool.setuptools.packages.find]
where = ["src"]

# ---------------------------------------------------------------------------
# Ruff – linting + formatting
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = [
    "E", "W",   # pycodestyle
    "F",        # pyflakes
    "I",        # isort
    "UP",       # pyupgrade
    "B",        # flake8-bugbear
    "C4",       # flake8-comprehensions
    "SIM",      # flake8-simplify
    "TCH",      # flake8-type-checking
    "ANN",      # flake8-annotations (enforce type hints)
    "RUF",      # ruff-specific rules
]
ignore = [
    "ANN401",  # allow Any in limited cases (will be caught by mypy)
    "TCH",     # first-party type-checking blocks — keep imports at top for clarity
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN"]
"scripts/**" = ["ANN"]

[tool.ruff.lint.isort]
known-first-party = ["alignai"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------------------------------------------------------------------
# Mypy – strict type checking
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = true
files = ["src/alignai", "scripts"]
exclude = ["tests/"]

# Third-party stubs
[[tool.mypy.overrides]]
module = [
    "trafilatura.*",
    "playwright.*",
    "docx.*",
    "telegram.*",
    "keyring.*",
    "import_linter.*",
]
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: end-to-end tests requiring real LLM endpoint (deselect with -m 'not integration')",
]
addopts = [
    "--strict-markers",
    "-q",
    "-m",
    "not integration",
]

[tool.coverage.run]
source = ["src/alignai"]
omit = ["src/alignai/ui/*", "src/alignai/main.py"]

[tool.coverage.report]
fail_under = 85
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@(abc\\.)?abstractmethod",
]

# ---------------------------------------------------------------------------
# Import-linter – enforce hexagonal layer isolation
# ---------------------------------------------------------------------------
[tool.importlinter]
root_packages = ["alignai"]
include_external_packages = true

[[tool.importlinter.contracts]]
name = "Domain must not import application, infra, agents, or ui"
type = "forbidden"
source_modules = ["alignai.domain"]
forbidden_modules = [
    "alignai.application",
    "alignai.infra",
    "alignai.agents",
    "alignai.ui",
]

[[tool.importlinter.contracts]]
name = "Application must not import infra, agents, or ui"
type = "forbidden"
source_modules = ["alignai.application"]
forbidden_modules = [
    "alignai.infra",
    "alignai.agents",
    "alignai.ui",
]

[[tool.importlinter.contracts]]
name = "Agents must not import ui"
type = "forbidden"
source_modules = ["alignai.agents"]
forbidden_modules = ["alignai.ui"]
